Use map to represent a  structure with named fields
Tuples can work well for some cases, but they have a problem:
fields aren’t named. You have to manually keep track of a tuple’s size
and the position of each field. This isn’t a problem for
abstractions that are unlikely to change, such as dates.
But tuples are less appropriate if the structure isn’t final and
is likely to change in the future. For such cases, you’re better
off using maps.

Using a map to represent a structure with named fields is simple.