Provide an alternative interface for update entry
Finally, note that this abstraction is still overly vague.
When clients provide entries via TodoList.add_entry/2 and
TodoList.update_entry/2, there are no restrictions on what the
entries should contain. These functions accept any kind of map.

To make this abstraction more restrictive, you could introduce a
dedicated struct for the entry (for example, a TodoEntry) and then
use pattern matching to enforce that each entry is in fact an instance
of that struct.