Limiting concurrency with pooling
Essentially, this technique keeps the concurrency level under control
and works best when dealing with resources that can’t cope with a large
number of concurrent requests.

The database process create three worker processes and keep their pids in
its internal state. When a request arrives, it’s delegated to one of the
worker processes it may be in a round-robin fashion or with some other
load-distribution strategy.